Refactor Program into directory
