+++ .gitignore
@@ -16,3 +16,4 @@ Cargo.lock
 msvc/
 gnu-mingw/
 SDL2.dll
+.DS_Store
