+++ src/gl/program/mod.rs
@@ -1,190 +1,17 @@
+pub mod questions;
+// mod errors;
+// mod shaders;
+// mod programs;
+
 use std::{ffi, ptr};
 use std::fmt;
 use std::error::Error;
 
-use error::{AppResult, AppError};
+use error::{AppResult};
 use gl;
 use gl::GlObject;
-use gl::error::GlError;
 use gl::raw::types::*;
 
-
-/// Asks Driver questions about the Given Program
-pub mod program_questions {
-    use super::*;
-    pub fn is_program(id: GLuint) -> AppResult<()> {
-        unsafe {
-            let result = gl::raw::IsProgram(id) as GLint;
-            if result == gl::raw::TRUE as GLint {
-                return Ok(());
-            } else if result == gl::raw::FALSE as GLint {
-                return Err(
-                    GlError::QuestionError(
-                        format!("Unknown error during Question `is_program`."),
-                    ).into(),
-                );
-            } else {
-                return Err(
-                    GlError::QuestionError(format!("{} is not referencing a glProgram", result))
-                        .into(),
-                );
-            }
-        }
-    }
-
-    pub fn is_linked(id: GLuint) -> AppResult<()> {
-        unsafe {
-            // Get the link status
-            let mut status = gl::raw::FALSE as GLint;
-            gl::raw::GetProgramiv(id, gl::raw::LINK_STATUS, &mut status);
-
-            // Fail on error
-            if status != (gl::raw::TRUE as GLint) {
-                let mut len = 0;
-                gl::raw::GetProgramiv(id, gl::raw::INFO_LOG_LENGTH, &mut len);
-                let mut buf: Vec<u8> = Vec::with_capacity(len as usize);
-                gl::raw::GetProgramInfoLog(id, len, ptr::null_mut(), buf.as_mut_ptr() as *mut GLchar);
-                buf.set_len(len as usize);
-                match String::from_utf8(buf) {
-                    Ok(s) => return Err(ProgramError::CompilationError(format!("{}", s)).into()),
-                    Err(e) => return Err(AppError::GenericError(format!("{:?}", e))),
-                }
-            }
-            Ok(())
-        }
-    }
-}
-
-/// Asks Driver questions about the Given Shader
-pub mod shader_questions {
-    use super::*;
-
-    /// Checks if given id is actually an shader, Errors if it is not
-    /// # Example
-    /// ```
-    /// use rustlike::context;
-    /// use rustlike::gl;
-    /// use rustlike::gl::GlObject;
-    /// use rustlike::gl::program::{self, ShaderKind};
-    /// use rustlike::gl::program::shader_questions;
-    /// use rustlike::gl::raw::types::*;
-    ///
-    /// let vertex_kind = ShaderKind::Vertex;
-    /// let ctx = context::ContextBuilder::default().build().unwrap();
-    /// let vertex_shader = program::CompiledShader::new(vertex_kind.example(), vertex_kind)
-    ///     .unwrap();
-    /// assert!(shader_questions::is_shader(vertex_shader.as_gl_id()).is_ok())
-    /// ```
-    ///
-    pub fn is_shader(id: GLuint) -> AppResult<()> {
-        unsafe {
-            let result = gl::raw::IsShader(id) as GLint;
-            if result == gl::raw::TRUE as GLint {
-                return Ok(());
-            } else if result == gl::raw::FALSE as GLint {
-                return Err(
-                    GlError::QuestionError(
-                        format!("Unknown error during Question `is_shader`."),
-                    ).into(),
-                );
-            } else {
-                return Err(
-                    GlError::QuestionError(format!("{} is not referencing a glShader", result))
-                        .into(),
-                );
-            }
-        }
-    }
-
-    /// Checks if Driver is marked for Deletion, Errors if it is not a shader
-    /// TODO: Example
-    pub fn is_deleted(id: GLuint) -> AppResult<bool> {
-        unsafe {
-            is_shader(id)?;
-            let mut status = gl::raw::TRUE as GLint;
-            gl::raw::GetShaderiv(id, gl::raw::DELETE_STATUS, &mut status);
-
-            if status == gl::raw::FALSE as GLint {
-                return Ok(false);
-            } else if status == gl::raw::TRUE as GLint {
-                return Ok(true);
-            }
-            return Err(
-                GlError::QuestionError(format!(
-                    "OpenGL Don Fucked up. Expect raw::TRUE or raw::FALSE, but got: {}",
-                    status
-                )).into(),
-            );
-        }
-    }
-
-    /// Returns kind of Shader in local abstraction, Errors if it is not a shader
-    /// # Example
-    /// ```
-    /// use rustlike::context;
-    /// use rustlike::gl;
-    /// use rustlike::gl::GlObject;
-    /// use rustlike::gl::program::{self, ShaderKind};
-    /// use rustlike::gl::program::shader_questions;
-    /// use rustlike::gl::raw::types::*;
-    ///
-    /// let fragment_kind = ShaderKind::Fragment;
-    /// let ctx = context::ContextBuilder::default().build().unwrap();
-    /// let fragment_shader = program::CompiledShader::new(fragment_kind.example(), fragment_kind)
-    ///     .unwrap();
-    /// assert_eq!(
-    ///     shader_questions::shader_kind(fragment_shader.as_gl_id()).unwrap(),
-    ///     fragment_kind
-    /// )
-    /// ```
-    ///
-    pub fn shader_kind(id: GLuint) -> AppResult<ShaderKind> {
-        unsafe {
-            is_shader(id)?;
-            let mut status = 0 as GLint;
-            gl::raw::GetShaderiv(id, gl::raw::SHADER_TYPE, &mut status);
-
-            if status == gl::raw::FRAGMENT_SHADER as GLint {
-                return Ok(ShaderKind::Fragment);
-            } else if status == gl::raw::VERTEX_SHADER as GLint {
-                return Ok(ShaderKind::Vertex);
-            }
-            return Err(
-                GlError::QuestionError(format!(
-                    "Expected Shader Type, from driver, but got: {}",
-                    status
-                )).into(),
-            );
-        }
-    }
-
-    pub fn is_successfully_compiled(id: GLuint) -> AppResult<()> {
-        unsafe {
-            let mut status = gl::raw::FALSE as GLint;
-            gl::raw::GetShaderiv(id, gl::raw::COMPILE_STATUS, &mut status);
-
-            if status != (gl::raw::TRUE as GLint) {
-                let mut len = 0;
-                gl::raw::GetShaderiv(id, gl::raw::INFO_LOG_LENGTH, &mut len);
-                let mut buf: Vec<u8> = Vec::with_capacity(len as usize);
-                gl::raw::GetShaderInfoLog(
-                    id,
-                    len,
-                    ptr::null_mut(),
-                    buf.as_mut_ptr() as *mut GLchar,
-                );
-
-                buf.set_len(len as usize);
-
-                match String::from_utf8(buf) {
-                    Ok(s) => return Err(ProgramError::CompilationError(format!("{}", s)).into()),
-                    Err(e) => return Err(AppError::GenericError(format!("{:?}", e))),
-                }
-            }
-            return Ok(());
-        }
-    }
-}
 
 /// ShaderKind is an typesafe representation of OpenGL shade types.
 /// These will map into `gl::raw` shader enum
@@ -284,7 +111,7 @@ impl CompiledShader {
             gl::raw::ShaderSource(glid, 1, &c_str.as_ptr(), ptr::null());
             gl::raw::CompileShader(glid);
 
-            shader_questions::is_successfully_compiled(glid)?;
+            questions::shader::is_successfully_compiled(glid)?;
             Ok(CompiledShader { glid, kind })
         }
     }
@@ -292,7 +119,7 @@ impl CompiledShader {
     /// Checks if the CompiledShader is an shader, has been successfully compiled,
     /// and isn't marked for deletion
     pub fn is_valid(&self) -> AppResult<()> {
-        let gpu_shader = shader_questions::shader_kind(self.glid)?;
+        let gpu_shader = questions::shader::shader_kind(self.glid)?;
         if gpu_shader != self.kind {
             return Err(
                 ProgramError::InvalidShader(format!(
@@ -302,8 +129,8 @@ impl CompiledShader {
                 )).into(),
             );
         }
-        shader_questions::is_deleted(self.glid)?;
-        shader_questions::is_successfully_compiled(self.glid)?;
+        questions::shader::is_deleted(self.glid)?;
+        questions::shader::is_successfully_compiled(self.glid)?;
         Ok(())
     }
 
@@ -344,7 +171,7 @@ impl ShaderProgram {
     /// ```
     /// use rustlike::context;
     /// use rustlike::gl::GlObject;
-    /// use rustlike::gl::program::{self, ShaderKind, program_questions};
+    /// use rustlike::gl::program::{self, ShaderKind, questions};
     ///
     /// let vertex_kind = ShaderKind::Vertex;
     /// let fragment_kind = ShaderKind::Fragment;
@@ -354,7 +181,7 @@ impl ShaderProgram {
     /// let fragment_shader = program::CompiledShader::new(fragment_kind.example(), fragment_kind)
     ///     .unwrap();
     /// let program = program::ShaderProgram::new(&vertex_shader, &fragment_shader).unwrap();
-    /// assert!(program_questions::is_program(program.as_gl_id()).is_ok());
+    /// assert!(questions::program::is_program(program.as_gl_id()).is_ok());
     /// ```
     pub fn new(
         vertex_shader: &CompiledShader,
@@ -385,7 +212,7 @@ impl ShaderProgram {
             gl::raw::AttachShader(program, vertex_shader.as_gl_id());
             gl::raw::AttachShader(program, fragment_shader.as_gl_id());
             gl::raw::LinkProgram(program);
-            program_questions::is_linked(program)?;
+            questions::program::is_linked(program)?;
             program
         };
